# High performance logging (NET6+)