--- a/src/SlimMessageBus.Host.Memory/Consumers/AbstractMessageProcessorQueue.cs
+++ b/src/SlimMessageBus.Host.Memory/Consumers/AbstractMessageProcessorQueue.cs
@@ -1,7 +1,9 @@
 ﻿namespace SlimMessageBus.Host.Memory;
 
-public abstract class AbstractMessageProcessorQueue(IMessageProcessor<object> messageProcessor, ILogger logger) : IMessageProcessorQueue
+public abstract partial class AbstractMessageProcessorQueue(IMessageProcessor<object> messageProcessor, ILogger logger) : IMessageProcessorQueue
 {
+    private readonly ILogger _logger = logger;
+
     public abstract void Enqueue(object transportMessage, IReadOnlyDictionary<string, object> messageHeaders);
 
     protected async Task ProcessMessage(object transportMessage, IReadOnlyDictionary<string, object> messageHeaders, CancellationToken cancellationToken)
@@ -23,7 +25,27 @@ public abstract class AbstractMessageProcessorQueue(IMessageProcessor<object> me
         if (r.Exception != null)
         {
             // We rely on the IMessageProcessor to execute the ConsumerErrorHandler<T>, but if it's not registered in the DI, it fails, or there is another fatal error then the message will be lost.
-            logger.LogError(r.Exception, "Error processing message {Message} of type {MessageType}", transportMessage, transportMessage.GetType());
+            LogMessageError(transportMessage, transportMessage.GetType(), r.Exception);
         }
     }
+
+    #region Logging
+
+    [LoggerMessage(
+       EventId = 0,
+       Level = LogLevel.Error,
+       Message = "Error processing message {TransportMessage} of type {TransportMessageType}")]
+    private partial void LogMessageError(object transportMessage, Type transportMessageType, Exception e);
+
+    #endregion
 }
+
+#if NETSTANDARD2_0
+
+public abstract partial class AbstractMessageProcessorQueue
+{
+    private partial void LogMessageError(object transportMessage, Type transportMessageType, Exception e)
+        => _logger.LogError(e, "Error processing message {TransportMessage} of type {TransportMessageType}", transportMessage, transportMessageType);
+}
+
+#endif