… src/SlimMessageBus.Host/Consumer/MessageProcessors/MessageProcessor.cs
… src/SlimMessageBus.Host/Consumer/MessageProcessors/ResponseMessageProcessor.cs
… src/SlimMessageBus.Host/Helpers/CompatRecord.cs
… src/SlimMessageBus.Host/Helpers/Utils.cs
… src/SlimMessageBus.Host/Hybrid/HybridMessageBus.cs
… src/SlimMessageBus.Host/MessageBusBase.cs
… src/SlimMessageBus.Host/RequestResponse/PendingRequestManager.cs
… src/SlimMessageBus.Host/Services/MessageHeaderService.cs
… src/SlimMessageBus/SlimMessageBus.csproj
… src/Tests/SlimMessageBus.Host.CircuitBreaker.Test/HealthCheckCircuitBreakerAbstractConsumerInterceptorTests.cs
… src/Tests/SlimMessageBus.Host.Memory.Benchmark/AbstractMemoryBenchmark.cs
… src/Tests/SlimMessageBus.Host.Memory.Benchmark/PubSubBenchmark.cs
… src/Tests/SlimMessageBus.Host.Memory.Benchmark/README.md
… src/Tests/SlimMessageBus.Host.Memory.Benchmark/ReqRespBenchmark.cs
… src/Tests/SlimMessageBus.Host.Memory.Test/MemoryMessageBusTests.cs
… src/Tests/SlimMessageBus.Host.Test/Consumer/AbstractConsumerTests.cs
… src/Tests/SlimMessageBus.Host.Test/Consumer/ResponseMessageProcessorTest.cs
… src/Tests/SlimMessageBus.Host.Test/Hybrid/HybridMessageBusTest.cs
… src/Tests/SlimMessageBus.Host.Test/ReqestResponse/PendingRequestStateTest.cs